[Commit] Iteration 934 #errors = 141 <= 140.
Addressing `BuildError `[/tmp/tempfspartition/QnetTransformArtifacts/MidTransformCode/sourceCode/BSTTiming-main/BSTTiming/Program.cs][CS0518][44,23] Predefined type 'System.Double' is not defined or imported``.

+++ BSTTiming-main/BSTTiming/Program.cs
@@ -71,7 +71,7 @@ using (System.IO.StreamWriter file = new System.IO.StreamWriter(@"C:\Users\Jesus
                 sw.Stop();
                 elapsed = msecs(sw);
             } while (elapsed < DURATION);
-            double totalAverage = elapsed / repetitions;
+            double totalAverage = elapsed / (double)repetitions;
 
             // Create a stopwatch
             sw = new Stopwatch();
